% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{create_profiles}
\alias{create_profiles}
\title{Use the histone reads to create profiles for a list of genomic sites}
\usage{
create_profiles(
  ranges,
  bam_file,
  reference,
  name = NA,
  bin_size = 100,
  ignore_strand = FALSE
)
}
\description{
Use the histone reads to create profiles for a list of genomic sites
}
