% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_enhancers.R
\name{find_enhancers}
\alias{find_enhancers}
\title{Find enhancer sites suitable to serve as training data.}
\usage{
find_enhancers(
  p300,
  DNase,
  window = 1000,
  N = NULL,
  TSS = NULL,
  min_dist_to_promoter = 2000,
  blacklist = NULL,
  verbose = TRUE
)
}
\description{
Find enhancer sites suitable to serve as training data.
}
