% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_promoters.R
\name{find_promoters}
\alias{find_promoters}
\title{Find promoter sites suitable for extracting a characteristic profile.}
\usage{
find_promoters(
  TSS_annotations,
  DNase,
  window = 1000,
  N = NULL,
  between_promoter_distance = 2000,
  blacklist = NULL,
  verbose = TRUE
)
}
\arguments{
\item{TSS_annotations}{A \code{GRanges} object containing the TSS annotations that will be used to
mark candidate promoter sites.}

\item{DNase}{A \code{GRanges} object containing DNase binding sites. There should be a
\code{peak} metadata column here, At least three metadata columns should be
present: \code{peak}, \code{qValue} and \code{signalValue}.}

\item{window}{The length of the window to cut around each promoter site. This window
can later be used to create the corresponding coverage profile using
\code{create_profiles}. The window will be centered on the promoter site.}

\item{N}{The number of promoter sites to find. Promoter sites with the highest
qValue will be used (using signalValue to resolve ties).}

\item{between_promoter_distance}{Desired minimum distance between the selected promoter sites.}

\item{blacklist}{An optional \code{GRanges} object containing areas that should be avoided
when searching for suitable promoter sites.}

\item{verbose}{A logical scalar indicating whether to print out informational messages
during the search for suitable promoter sites.}
}
\value{
\preformatted{A `GRanges` object containing windows centered around suitable promoter
sites. The promoter site itself lies in the center of the window. The
`RGanges` object has a `type` metadata column that is set to
`'promoter'`.
}
}
\description{
Finds promoter sites suitable to serve as training data for building a
characteristic profile for use with the pattern_likelihoods function.
}
\details{
Promoter sites are sites on the genome that have TSS annotations and are
also DNase binding sites.
}
\seealso{
\code{\link[=find_enhancers]{find_enhancers()}} for finding suitable enhancer sites and
\code{\link[=find_random]{find_random()}} for finding suitable random sites.
}
