% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_random.R
\name{find_random}
\alias{find_random}
\title{Find random sites in the genome that are not enhancers or promoters.}
\usage{
find_random(
  window = 1000,
  N = NULL,
  chroms_of_interest = NULL,
  p300 = NULL,
  max_dist_to_p300 = 2000,
  TSS = NULL,
  max_dist_to_TSS = 2000,
  blacklist = NULL,
  verbose = TRUE
)
}
\description{
Find random sites in the genome that are not enhancers or promoters.
}
