% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{normalize_profiles}
\alias{normalize_profiles}
\title{Normalize one set of profiles with another}
\usage{
normalize_profiles(profiles, reference)
}
\description{
Normalize one set of profiles with another
}
