% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pattern_likelihoods}
\alias{pattern_likelihoods}
\title{Meat of the PREPRINT method.}
\usage{
pattern_likelihoods(
  profiles,
  agg_patterns,
  measure = "ML",
  positive_class = "enhancer"
)
}
\description{
Meat of the PREPRINT method.
}
