% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS_protein_coding.R
\name{TSS_protein_coding}
\alias{TSS_protein_coding}
\title{Use the histone reads to create profiles for a list of genomic sites}
\usage{
TSS_protein_coding(gencode_tss_file, verbose = TRUE)
}
\description{
Use the histone reads to create profiles for a list of genomic sites
}
