% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{create_profiles}
\alias{create_profiles}
\title{Create profiles for a list of genome ranges, using the genome coverage from a BAM file.}
\usage{
create_profiles(
  ranges,
  bam_file,
  reference,
  name = NA,
  bin_size = 100,
  ignore_strand = FALSE
)
}
\arguments{
\item{ranges}{The ranges for which to make profiles. These should be windows of a
fixed size, centered on a site of interest. These ranges can for example
be created by the find_enhancers, find_promoters or find_random
functions.}

\item{bam_file}{The BAM File to read the coverage data from.}

\item{reference}{The profiles to use as reference. The coverage counts from the reference
are subtracted from the profile. Can be NULL to specify no reference.}

\item{name}{A name for the features which coverage we are computing. This is used to
build the colnames() for the profiles. Defaults to NA, meaning the name
will be automatically inferred from the BAM file.}

\item{bin_size}{The width of the bins, i.e. the resolution of the profiles.}

\item{ignore_strand}{Whether to ignore the strand when computing coverage.}
}
\value{
The profiles.
}
\description{
Create profiles for a list of genome ranges, using the genome coverage from a BAM file.
}
