% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_promoters.R
\name{find_promoters}
\alias{find_promoters}
\title{Find promoter sites suitable to serve as training data.}
\usage{
find_promoters(
  TSS_annotation,
  DNase,
  window = 1000,
  N = NULL,
  between_TSS_distance = 2000,
  blacklist = NULL,
  verbose = TRUE
)
}
\description{
Find promoter sites suitable to serve as training data.
}
