% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{pattern_likelihoods}
\alias{pattern_likelihoods}
\title{Compute for a set of profiles, the likelihood of them being an enhancer site.}
\usage{
pattern_likelihoods(
  profiles,
  characteristic_profiles,
  measure = "Bayesian",
  positive_class = "enhancer"
)
}
\arguments{
\item{profiles}{The profiles to compute the likelihoods for.}

\item{characteristic_profiles}{The characteristic profiles for each site type (enhancer, promoter, random, etc.)}

\item{measure}{The likelihood measure to use. Can be either "Bayesian" for Bayesian
likelihood (the default), or "ML" for maximum likelihood.}

\item{positive_class}{When using Bayesian likelihood, you need to specify the site type
("enhancer", "promoter", "random", etc.) that is regarded as the
positive class, i.e. the type you are looking to detect.
Defaults to "enhancer".}
}
\value{
likelihoods
A copy of the given Profiles object, but with the profile data replaced
with likelihood scores. For each profile, likelihood scores are given
for each of the given characteristic profiles. The score reflects the
likelihood that the profile is of the type described by the
characteristic profile.
}
\description{
Checks each profile against a list of characteristic profiles. For each
characteristic profile, a likelihood score is computed that reflects how
likely it is that the profile is of the type (enhancer, promoter, random,
etc.) represented by the characteristic profile. This score can be computed
either in Bayesian fashion (which is the main method advocated by PREPRINT)
or maximum-likelihood.
}
